moved animations to showcase
