+++ src/app/docs/animations/hovers/jiggle/page.mdx
@@ -1,7 +1,7 @@
 import Link from 'next/link'
 import { RequiresTailwind } from '@/components/Requires'
-import HoverAnimations from '@/components/HoverAnimationCards'
-import { HoverJiggleButton } from '@/components/PreviewCode/HoverAnimationButtons'
+import HoverAnimations from '@/components/Cards/HoverAnimationCards'
+import { HoverJiggleButton } from '@/showcase/animations/HoverAnimationButtons'
 import { ChevronLeft } from 'lucide-react'
 import PreviewCodeTabs from '@/components/CodePreview/page'
