+++ src/app/docs/animations/skewed-infinite-scroll/page.mdx
@@ -1,4 +1,4 @@
-import SkewedInfiniteScroll from '@/ui/animation/SkewedInfiniteScroll'
+import SkewedInfiniteScroll from '@/showcase/animations/SkewedInfiniteScroll'
 import { CodeGroup } from '@/components/Code'
 import { RequiresTailwind } from '@/components/Requires'
 import {
